refactor: extract PropertyComparer & PropertyValueExtractor
