+++ src/ByteSync.Client/Interfaces/Services/Filtering/IPropertyComparer.cs
@@ -1,0 +1,9 @@
+﻿using ByteSync.Business.Filtering;
+using ByteSync.Business.Filtering.Values;
+
+namespace ByteSync.Interfaces.Services.Filtering;
+
+public interface IPropertyComparer
+{
+    bool CompareValues(PropertyValueCollection collection1, PropertyValueCollection collection2, FilterOperator filterOperator);
+}
